refactor: add external links constants
