--- src/constants/apiConstants.ts
@@ -1,1 +1,0 @@
-export const API_BASE_URL = 'http://109.68.212.118/';
