+++ src/constants/externalLinks.ts
@@ -1,0 +1,3 @@
+export const API_BASE_URL = 'http://109.68.212.118/';
+
+export const LAB_PROJECT_URL = 'https://profeaturelab.ru/';
